增加默认配置

+++ get4j-core/src/main/java/com/bytegriffin/get4j/annotation/Cascade.java
@@ -21,9 +21,9 @@ public @interface Cascade {
 
     String startTime() default "";
 
-    long interval() default 0;
+    int interval() default 0;
 
-    long sleep() default 0;
+    int sleep() default 0;
 
     String sleepRange() default "";
