+++ get4j-core/src/main/java/com/bytegriffin/get4j/annotation/ListDetail.java
@@ -26,9 +26,9 @@ public @interface ListDetail {
 
     String startTime() default "";
 
-    long interval() default 0;
+    int interval() default 0;
 
-    long sleep() default 0;
+    int sleep() default 0;
 
     String sleepRange() default "";
