+++ get4j-core/src/main/java/com/bytegriffin/get4j/net/http/SleepRandomSelector.java
@@ -5,17 +5,17 @@ import java.util.List;
 /**
  * 每次请求的时间间隔随机选择器
  */
-public class SleepRandomSelector extends RandomSelector<Long> {
+public class SleepRandomSelector extends RandomSelector<Integer> {
 
-    private List<Long> queue;
+    private List<Integer> queue;
 
     @Override
-    protected List<Long> getQueue() {
+    protected List<Integer> getQueue() {
         return queue;
     }
 
     @Override
-    public void setQueue(List<Long> queue) {
+    public void setQueue(List<Integer> queue) {
         this.queue = queue;
     }
